Fix vite compile issue - vitest import in frontend code.
