+++ src/web/package.json
@@ -73,17 +73,10 @@
     "xml-formatter": "^2.6.1"
   },
   "devDependencies": {
-    "@snowpack/plugin-dotenv": "^2.1.0",
-    "@snowpack/plugin-postcss": "^1.4.3",
-    "@snowpack/plugin-react-refresh": "^2.5.0",
-    "@snowpack/plugin-sass": "^1.4.0",
-    "@snowpack/plugin-typescript": "^1.2.1",
-    "@snowpack/plugin-webpack": "^3.0.0",
     "@types/node": "^17.0.40",
     "@types/react": "^17.0.4",
     "@types/react-dom": "^17.0.3",
     "@types/react-modal": "^3.13.1",
-    "@types/snowpack-env": "^2.3.3",
     "@vitejs/plugin-react": "^1.3.2",
     "autoprefixer": "^10.2.5",
     "concurrently": "^7.2.1",
@@ -92,8 +85,6 @@
     "npm-run-all": "^4.1.5",
     "prettier": "^2.2.1",
     "sass": "1.52.2",
-    "snowpack": "^3.8.6",
-    "snowpack-plugin-raw-file-loader": "^1.0.9",
     "tsup": "^6.1.2",
     "typescript": "^4.7.2",
     "vite": "^2.9.12",
