+++ src/web/src/browser/presenter/index.ts
@@ -1,6 +1,4 @@
-import { mock } from 'vitest-mock-extended';
-import { createOvermind, createOvermindMock, IContext } from 'overmind';
-import { vi } from 'vitest';
+import { createOvermind, IContext } from 'overmind';
 
 import type { AnnotateXMLUseCase } from '@asap/shared/use-cases/annotate-xml';
 import type { AppMetrics } from '@asap/shared/use-cases/app-metrics';
@@ -13,7 +11,7 @@ import * as actions from './actions';
 import type { Location } from './state/router';
 import { state, State, SampleDocument } from './state';
 
-type UseCases = {
+export type UseCases = {
   annotateXML: AnnotateXMLUseCase;
   getAssertionViews: GetAssertionViews;
   getSchematronAssertions: GetSchematronAssertions;
@@ -67,23 +65,3 @@ export const createPresenter = (ctx: PresenterContext) => {
   return presenter;
 };
 export type Presenter = ReturnType<typeof createPresenter>;
-
-type MockPresenterContext = {
-  useCases?: Partial<UseCases>;
-  initialState?: Partial<State>;
-};
-
-export const createPresenterMock = (ctx: MockPresenterContext = {}) => {
-  const presenter = createOvermindMock(
-    getPresenterConfig(
-      { getCurrent: vi.fn(), listen: vi.fn(), replace: vi.fn() },
-      mock<UseCases>(),
-      ctx.initialState,
-    ),
-    {
-      useCases: ctx.useCases,
-    },
-  );
-  return presenter;
-};
-export type PresenterMock = ReturnType<typeof createPresenterMock>;
